requirements.txt
appdirs==1.4.3
appnope==0.1.0
asn1crypto==1.2.0
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.9.0
bleach==3.1.1
blis==0.4.1
Bottleneck==1.3.2
catalogue==1.0.0
certifi==2019.11.28
cffi==1.13.0
chardet==3.0.4
click==7.1.1
combo==0.1.0
conda==4.8.2
conda-package-handling==1.6.0
cryptography==2.8
cycler==0.10.0
cymem==2.0.3
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
entrypoints==0.3
env==0.1.0
fastai==1.0.60
fastprogress==0.2.3
filelock==3.0.12
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
helper==2.4.2
idna==2.8
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.0.0
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.1.0
llvmlite==0.31.0
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numba==0.48.0
numexpr==2.7.1
numpy==1.18.1
nvidia-ml-py3==7.352.0
packaging==20.3
pandas==1.0.2
pandocfilters==1.4.2
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.1
plac==1.1.3
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==3.0.4
ptyprocess==0.6.0
pycosat==0.6.3
pycparser==2.19
Pygments==2.6.1
pynvx==1.0.0
pyod==0.7.8
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyrsistent==0.15.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
pyzmq==19.0.0
qtconsole==4.7.1
QtPy==1.9.0
requests==2.22.0
ruamel-yaml==0.15.46
scikit-learn==0.22.2.post1
scipy==1.4.1
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
soupsieve==2.0
spacy==2.2.4
SQLAlchemy==1.3.16
srsly==1.0.2
suod==0.0.3
terminado==0.8.3
testpath==0.4.4
thinc==7.4.0
torch==1.4.0
torchvision==0.5.0
tornado==6.0.4
tqdm==4.45.0
traitlets==4.3.3
urllib3==1.24.2
virtualenv==20.0.20
wasabi==0.6.0
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
zipp==3.1.0
